+++ library/src/test/java/org/mustangproject/ZUGFeRD/ZF2ZInvoiceImporterTest.java
@@ -23,6 +23,7 @@ package org.mustangproject.ZUGFeRD;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.Test;
 import org.mustangproject.*;
 
 import javax.xml.xpath.XPathExpressionException;
@@ -35,6 +36,9 @@ import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Arrays;
 import java.util.Date;
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 
 /***
@@ -369,6 +373,26 @@ public class ZF2ZInvoiceImporterTest extends ResourceCase {
 			throw new RuntimeException(e);
 		}
 
+
+	}
+
+	@Test
+	public void testImportIncludedNotes() throws XPathExpressionException, ParseException {
+		InputStream inputStream = this.getClass()
+			.getResourceAsStream("/EN16931_Einfach.pdf");
+		ZUGFeRDInvoiceImporter importer = new ZUGFeRDInvoiceImporter(inputStream);
+		Invoice invoice = importer.extractInvoice();
+		List<IncludedNote> notesWithSubjectCode = invoice.getNotesWithSubjectCode();
+		assertThat(notesWithSubjectCode).hasSize(2);
+		assertThat(notesWithSubjectCode.get(0).getSubjectCode()).isNull();
+		assertThat(notesWithSubjectCode.get(0).getContent()).isEqualTo("Rechnung gemäß Bestellung vom 01.11.2024.");
+		assertThat(notesWithSubjectCode.get(1).getSubjectCode()).isEqualTo(SubjectCode.REG);
+		assertThat(notesWithSubjectCode.get(1).getContent()).isEqualTo("Lieferant GmbH\t\t\t\t\n"
+			+ "Lieferantenstraße 20\t\t\t\t\n"
+			+ "80333 München\t\t\t\t\n"
+			+ "Deutschland\t\t\t\t\n"
+			+ "Geschäftsführer: Hans Muster\n"
+			+ "Handelsregisternummer: H A 123");
 
 	}