+++ History.md
@@ -15,7 +15,7 @@
 - 538 Mustang validator always claims PDF is invalid if flavour is PDF/A-3A
 - 555 be able to validate ubl credit notes
 - when parsing now distinguishing between the parseExceptions StructureException and ArithmetricException
-
+- Import IncludedNotes on invoice extraction #554
 
 2.14.2
 =======